refactor: 디렉터리명 수정 및 UI 안정성 보완 (#9)
